Don't be concerned with GCP-specific things in PollResultMonitorActor
