+++ supportedBackends/google/batch/src/main/scala/cromwell/backend/google/batch/actors/BatchPollResultMonitorActor.scala
@@ -1,6 +1,7 @@
 package cromwell.backend.google.batch.actors
 
 import akka.actor.{ActorRef, Props}
+import cats.data.Validated.{Invalid, Valid}
 import cromwell.backend.{BackendJobDescriptor, BackendWorkflowDescriptor, Platform}
 import cromwell.backend.google.batch.models.RunStatus
 import cromwell.backend.standard.pollmonitoring.{
@@ -12,7 +13,7 @@ import cromwell.backend.standard.pollmonitoring.{
 }
 import cromwell.backend.validation.ValidatedRuntimeAttributes
 import cromwell.core.logging.JobLogger
-import cromwell.services.cost.InstantiatedVmInfo
+import cromwell.services.cost.{GcpCostLookupRequest, GcpCostLookupResponse, InstantiatedVmInfo}
 import cromwell.services.metadata.CallMetadataKeys
 
 import java.time.OffsetDateTime
@@ -44,6 +45,27 @@ class BatchPollResultMonitorActor(pollMonitorParameters: PollMonitorParameters)
   override def extractEndTimeFromRunState(pollStatus: RunStatus): Option[OffsetDateTime] =
     pollStatus.eventList.collectFirst {
       case event if event.name == CallMetadataKeys.VmEndTime => event.offsetDateTime
+    }
+
+  override def handleVmCostLookup(vmInfo: InstantiatedVmInfo) = {
+    val request = GcpCostLookupRequest(vmInfo, self)
+    params.serviceRegistry ! request
+  }
+
+  def handleCostResponse(costLookupResponse: GcpCostLookupResponse): Unit =
+    if (vmCostPerHour.isEmpty) { // Optimization to avoid processing responses after we've received a valid one.
+      val cost = costLookupResponse.calculatedCost match {
+        case Valid(c) =>
+          params.logger.info(s"vmCostPerHour for ${costLookupResponse.vmInfo} is ${c}")
+          c
+        case Invalid(errors) =>
+          params.logger.error(
+            s"Failed to calculate VM cost per hour for ${costLookupResponse.vmInfo}. ${errors.toList.mkString(", ")}"
+          )
+          BigDecimal(-1)
+      }
+      vmCostPerHour = Option(cost)
+      tellMetadata(Map(CallMetadataKeys.VmCostPerHour -> cost))
     }
 
   override def receive: Receive = {
