+++ supportedBackends/google/pipelines/common/src/main/scala/cromwell/backend/google/pipelines/common/PapiPollResultMonitorActor.scala
@@ -1,12 +1,13 @@
 package cromwell.backend.google.pipelines.common
 
 import akka.actor.{ActorRef, Props}
+import cats.data.Validated.{Invalid, Valid}
 import cromwell.backend.google.pipelines.common.api.RunStatus
 import cromwell.backend.standard.pollmonitoring._
 import cromwell.backend.validation.ValidatedRuntimeAttributes
 import cromwell.backend.{BackendJobDescriptor, BackendWorkflowDescriptor, Platform}
 import cromwell.core.logging.JobLogger
-import cromwell.services.cost.{GcpCostLookupResponse, InstantiatedVmInfo}
+import cromwell.services.cost.{GcpCostLookupRequest, GcpCostLookupResponse, InstantiatedVmInfo}
 import cromwell.services.metadata.CallMetadataKeys
 
 import java.time.OffsetDateTime
@@ -29,6 +30,7 @@ class PapiPollResultMonitorActor(parameters: PollMonitorParameters) extends Poll
 
   override def extractEarliestEventTimeFromRunState(pollStatus: RunStatus): Option[OffsetDateTime] =
     pollStatus.eventList.minByOption(_.offsetDateTime).map(e => e.offsetDateTime)
+
   override def extractStartTimeFromRunState(pollStatus: RunStatus): Option[OffsetDateTime] =
     pollStatus.eventList.collectFirst {
       case event if event.name == CallMetadataKeys.VmStartTime => event.offsetDateTime
@@ -42,7 +44,28 @@ class PapiPollResultMonitorActor(parameters: PollMonitorParameters) extends Poll
   override def extractVmInfoFromRunState(pollStatus: RunStatus): Option[InstantiatedVmInfo] =
     pollStatus.instantiatedVmInfo
 
+  override def handleVmCostLookup(vmInfo: InstantiatedVmInfo) = {
+    val request = GcpCostLookupRequest(vmInfo, self)
+    params.serviceRegistry ! request
+  }
+
   override def params: PollMonitorParameters = parameters
+
+  def handleCostResponse(costLookupResponse: GcpCostLookupResponse): Unit =
+    if (vmCostPerHour.isEmpty) { // Optimization to avoid processing responses after we've received a valid one.
+      val cost = costLookupResponse.calculatedCost match {
+        case Valid(c) =>
+          params.logger.info(s"vmCostPerHour for ${costLookupResponse.vmInfo} is ${c}")
+          c
+        case Invalid(errors) =>
+          params.logger.error(
+            s"Failed to calculate VM cost per hour for ${costLookupResponse.vmInfo}. ${errors.toList.mkString(", ")}"
+          )
+          BigDecimal(-1)
+      }
+      vmCostPerHour = Option(cost)
+      tellMetadata(Map(CallMetadataKeys.VmCostPerHour -> cost))
+    }
 
   override def receive: Receive = {
     case costResponse: GcpCostLookupResponse => handleCostResponse(costResponse)
@@ -53,13 +76,11 @@ class PapiPollResultMonitorActor(parameters: PollMonitorParameters) extends Poll
           params.logger.error(
             s"Programmer error: Received Poll Result of unknown type. Expected ${RunStatus.getClass.getSimpleName} but got ${result.getClass.getSimpleName}."
           )
-
         case AsyncJobHasFinished(pollResult: RunStatus) => handleAsyncJobFinish(pollResult.getClass.getSimpleName)
         case AsyncJobHasFinished(result) =>
           params.logger.error(
             s"Programmer error: Received Poll Result of unknown type. Expected ${AsyncJobHasFinished.getClass.getSimpleName} but got ${result.getClass.getSimpleName}."
           )
-
       }
     case unexpected =>
       params.logger.error(
